Koenvandesande remove duplicate filenames (#448)
* Remove duplicate filenames which do not work on Windows by merging files

* Fix

* relu tests
