--- torch2trt/converters/ReLU.py
@@ -1,11 +1,0 @@
-from torch2trt.torch2trt import *
-
-
-@tensorrt_converter('torch.nn.ReLU.forward')
-def convert_ReLU(ctx):
-    input = ctx.method_args[1]
-    input_trt = add_missing_trt_tensors(ctx.network, [input])[0]
-    output = ctx.method_return
-    layer = ctx.network.add_activation(
-        input=input_trt, type=trt.ActivationType.RELU)
-    output._trt = layer.get_output(0)
