+++ torch2trt/converters/__init__.py
@@ -12,11 +12,8 @@ from .Conv1d import *
 from .Conv2d import *
 from .ConvTranspose import *
 from .ConvTranspose2d import *
-from .Identity import *
 from .Linear import *
 from .LogSoftmax import *
-from .ReLU import *
-from .ReLU6 import *
 from .activation import *
 from .adaptive_avg_pool2d import *
 from .adaptive_max_pool2d import *
